Make neovim healthchecks warn about one less thing
Even though we manage the neovim configuration solely with nix, the
neovim health checker warns that we're not doing things right since
there is no such file at `~/.config/nvim/init.vim`. It's just a warning
but it's something easily remedied and can point people to the right
spot to actually edit the neovim configuration.